feat(emails): add mock data for non-production environments
This commit introduces a mock data object for the email template. This
object is used when the NODE_ENV is not 'production', allowing for
easier testing and development. The email template now uses this data
object instead of props directly. Additionally, the email template
layout has been slightly adjusted and package versions have been
updated.